chore: Extract debugger code to doctest/parts/debugger.{h,cpp}
