+++ CMakeLists.txt
@@ -63,7 +63,11 @@ endif()
 ################################################################################
 
 if(${DOCTEST_WITH_MAIN_IN_STATIC_LIB})
-    add_library(${PROJECT_NAME}_with_main STATIC EXCLUDE_FROM_ALL ${doctest_parts_folder}/private/doctest.cpp)
+    add_library(
+        ${PROJECT_NAME}_with_main STATIC EXCLUDE_FROM_ALL
+        ${doctest_parts_folder}/private/doctest.cpp
+        ${doctest_parts_folder}/private/debugger.cpp
+    )
     add_library(${PROJECT_NAME}::${PROJECT_NAME}_with_main ALIAS ${PROJECT_NAME}_with_main)
     target_compile_definitions(${PROJECT_NAME}_with_main PRIVATE
         DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN)
